Split Client.h out of PluginManager.h to fix layering.
